refactor: 패키지 이름을 mapper로 변경
