+++ src/main/java/racingcar/mapper/PlayResultDao.java
@@ -1,4 +1,4 @@
-package racingcar.dao;
+package racingcar.mapper;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
